+++ CHANGELOG.md
@@ -1,13 +1,23 @@
 # CHANGELOG
 
+## Upcoming release
+
+- Device identifiers become long-lived
+- `launch()`, `unlaunch()`, `enableRemoteNotifications()` and `disableRemoteNotifications()` become suspending functions with a callback alternative
+
+#### Breaking changes
+
+- `NotificareDevice.id` attribute no longer contains the push token. Use `Notificare.push().subscriptionId` instead.
+- The `NotificareDevice` data model was reduced to only publicly relevant attributes.
+- `didRegisterDevice` only triggers once, when the device is created.
+- `launch()`, `unlaunch()`, `enableRemoteNotifications()` and `disableRemoteNotifications()` become suspending functions that complete after all the work is done.
+- `NotificareTransport` was moved to the push module.
+- Drops support for the monetize module.
+
 ## 3.9.1
 
 - Preload images before showing in-app messages
 - Fix `NotificarePass.type` decoding
-
-#### Breaking changes
-
-- Drops support for the monetize module.
 
 ## 3.9.0